+++ client/src/types.ts
@@ -1,0 +1,10 @@
+export type Subscription = {
+  _id?: number;
+  icon:string;
+  price: number;
+  title: string;
+  start: string;
+  prettyStart:string;
+  cycle: string;
+  reminderDate: string;
+}
